examples/kitt/requirements.txt
livekit-agents~=0.5.dev0
livekit-plugins-deepgram~=0.3.dev0
livekit-plugins-elevenlabs~=0.3.dev0
livekit-plugins-openai~=0.3.dev0
python-dotenv
azure-cognitiveservices-speech